Refactored lightning example
