+++ integrations-and-supported-tools/pytorch-lightning/scripts/requirements.txt
@@ -1,7 +1,6 @@
 lightning
 neptune
 pydantic
-scikit-plot
 scipy<1.12
 torch
 torchvision
