+++ integrations-and-supported-tools/pytorch-lightning/scripts/run_examples.sh
@@ -1,10 +1,7 @@
 set -e
 
 echo "Installing requirements..."
-pip install -U -r requirements.txt
+pip install -q -U -r requirements.txt
 
-echo "Running Neptune_Pytorch_Support.py..."
+echo "Running Neptune_Pytorch_Lightning.py..."
 python Neptune_Pytorch_Lightning.py
-
-echo "Running Neptune_Pytorch_Support_more_options.py..."
-python Neptune_Pytorch_Lightning_more_options.py
